+++ packages/image/src/index.tsx
@@ -1,6 +1,9 @@
 import Image from './components/Image'
 import ImageProvider from './components/ImageProvider'
+import useImage from './hooks/useImage'
+import getImageDataURL from './methods/getImageDataURL'
+import preloadImage from './methods/preloadImage'
 
-export { ImageProvider }
+export { ImageProvider, useImage, getImageDataURL, preloadImage }
 
 export default Image