+++ packages/image/src/interfaces/hooks/useImageOptions.ts
@@ -1,0 +1,5 @@
+import ImageCacheOptions from '../ImageCacheOptions'
+
+export default interface useImageOptions {
+  cache?: ImageCacheOptions
+}