+++ packages/image/src/methods/getImageClassName.ts
@@ -1,0 +1,7 @@
+export default function getImageClassName (fill?: boolean) {
+  let className = ''
+  if (fill) {
+    className += 'w-full h-full'
+  }
+  return className
+}