+++ packages/image/src/methods/getImageDataURL.ts
@@ -1,0 +1,7 @@
+export default async function getImageDataURL (blob: Blob | File) {
+  return await new Promise<string>((resolve) => {
+    const reader = new FileReader()
+    reader.onloadend = () => resolve(reader.result as string)
+    reader.readAsDataURL(blob)
+  })
+}